+++ src/backend/resolvers/vega/mutations/personRole/delete.ts
@@ -3,7 +3,7 @@ import axios from 'axios'
 const deleteVegaPersonRole = async (
   parent,
   { input: { id } },
-  context,
+  { pulseDevDb },
   info
 ) => {
   const deleteVegaPersonRole = await axios.get(`people-roles/${id}/`)
@@ -13,6 +13,20 @@ const deleteVegaPersonRole = async (
     .catch((e) => {
       throw new Error(JSON.stringify(e.response.data))
     })
+
+  await pulseDevDb.collection('marketBasketsSurveyAnswers').updateMany(
+    { 'stakeholder.primaryRoleId': id },
+    {
+      $set: {
+        'stakeholder.primaryRoleId': null,
+        'stakeholder.primaryRole': null,
+        'stakeholder.primaryRoleTypeId': null,
+        'stakeholder.primaryRoleType': null,
+        'stakeholder.roleSpecialtyId': null,
+        'stakeholder.roleSpecialty': null,
+      }
+    }
+  )
 
   return deleteVegaPersonRole
 }