+++ src/backend/resolvers/vega/mutations/personRole/indication/update.ts
@@ -1,16 +1,27 @@
 import axios from 'axios'
 
-const updateVegaPersonRoleIndication = (
+const updateVegaPersonRoleIndication = async (
   parent,
   { input: { id, ...body } },
-  context,
+  { pulseDevDb },
   info
 ) => {
-  return axios.patch(`people-roles-indications/${id}/`, body)
+  const updatedRoleSpecialty = await axios.patch(`people-roles-indications/${id}/`, body)
     .then(({ data }) => data)
     .catch((e) => {
       throw new Error(JSON.stringify(e.response.data))
     })
+
+  await pulseDevDb.collection('marketBasketsSurveyAnswers').updateMany(
+    { 'stakeholder.roleSpecialtyId': id },
+    {
+      $set: {
+        'stakeholder.roleSpecialty': updatedRoleSpecialty.specialty_label,
+      }
+    }
+  )
+
+  return updatedRoleSpecialty
 }
 
 export default updateVegaPersonRoleIndication