+++ src/backend/resolvers/vega/mutations/personRole/update.ts
@@ -3,14 +3,27 @@ import axios from 'axios'
 const updateVegaPersonRole = async (
   parent,
   { input: { id, ...body } },
-  context,
+  { pulseDevDb },
   info
 ) => {
-  return axios.patch(`people-roles/${id}/`, body)
+  const updatedRole = await axios.patch(`people-roles/${id}/`, body)
     .then(({ data }) => data)
     .catch((e) => {
       throw new Error(JSON.stringify(e.response.data))
     })
+
+  await pulseDevDb.collection('marketBasketsSurveyAnswers').updateMany(
+    { 'stakeholder.primaryRoleId': id },
+    {
+      $set: {
+        'stakeholder.primaryRole': updatedRole.name,
+        'stakeholder.primaryRoleTypeId': updatedRole.type && updatedRole.type.id,
+        'stakeholder.primaryRoleType': updatedRole.type && updatedRole.type.name,
+      }
+    }
+  )
+
+  return updatedRole
 }
 
 export default updateVegaPersonRole