+++ src/backend/resolvers/vega/mutations/personRoleType/delete.ts
@@ -3,7 +3,7 @@ import axios from 'axios'
 const deleteVegaPersonRoleType = async (
   parent,
   { input: { id } },
-  context,
+  { pulseDevDb },
   info
 ) => {
   const deleteVegaPersonRoleType = await axios.get(`people-roles-types/${id}/`)
@@ -13,6 +13,16 @@ const deleteVegaPersonRoleType = async (
     .catch((e) => {
       throw new Error(JSON.stringify(e.response.data))
     })
+
+  await pulseDevDb.collection('marketBasketsSurveyAnswers').updateMany(
+    { 'stakeholder.primaryRoleTypeId': id },
+    {
+      $set: {
+        'stakeholder.primaryRoleTypeId': null,
+        'stakeholder.primaryRoleType': null,
+      }
+    }
+  )
 
   return deleteVegaPersonRoleType
 }